Enabling network isolation of BOSH services
1. Initial implementation done for OpenStack alone.
2. Isolation level is at CF space level. This is done keeping in
view of the limits on number of security groups per VPC/Tenant.
3. As part of service creation, first a check is done to see
if a security group with that space guid exists. If not, then
a security group with that space guid is created. Additionally
cloud config is also created with a network definition, reflecting
the newly created security group which is then referred in services
manifest.